mwan/go: align lint config with clyde-dev set; fix all 33 issues
Schema: switch from `extends:` (not a real golangci-lint feature; was
silently ignored, leaving us on default 5 linters) to inlined Clyde
linter set: bodyclose, copyloopvar, errcheck, errorlint, gocritic, govet,
ineffassign, intrange, misspell, modernize, nilerr, staticcheck,
unconvert, unparam, unused, usestdlibvars, usetesting, wastedassign,
whitespace + gofumpt/goimports formatters. Pairs with go-makefile@60c6991
which aligned the upstream template.

Add `make fmt` and `make fix` targets for `golangci-lint fmt` and
`golangci-lint run --fix`.

Issues resolved (33 -> 0):
* gofumpt (10) + modernize (9) + intrange (3) + errorlint (2): auto-fixed
* defer Body.Close errcheck (3): wrapped in `func() { _ = ... }()`
* unconvert (1): drop redundant int64() on already-int64
* nilerr (1): pveapi/client.go decodeAgentB64 was returning nil on
  base64 decode failure when the input had non-printable bytes,
  masking corruption. Surface the error properly with %w.
* unparam (7): mostly mechanical (unused params -> _) with two
  intentional cases marked nolint:
  - cmdDisarmWatchdog returns always-nil error to match dispatch sig
  - runCmd `name` always "qm" today but kept generic for non-qm usage

+++ mwan/go/internal/bgp/speaker.go
@@ -388,7 +388,7 @@ func (s *Speaker) Status() Status {
 		if p.State.SessionState == apipb.PeerState_SESSION_STATE_ESTABLISHED {
 			ps.Established = true
 			if p.Timers != nil && p.Timers.State != nil {
-				ps.UpSince = int64(p.Timers.State.Uptime.GetSeconds())
+				ps.UpSince = p.Timers.State.Uptime.GetSeconds()
 			}
 		}
 
@@ -407,7 +407,6 @@ func (s *Speaker) Status() Status {
 
 		st.Peers = append(st.Peers, ps)
 	})
-
 	if err != nil {
 		s.log.Error("list peers failed", "error", err)
 	}
